Implement game flow for Scrabbling Game
When the player launches the game, it will randomly load a word, initialise the tile bag of characters, and display a text-based board. 
The game will then validate dictionary word list from https://raw.githubusercontent.com/mycourseresource/datastore/master/computingterms.yaml with a minimum of 5 characters, AND characters on the tile bag should not exceed a maximum of 12 characters using logical operators in character length.
The game will prompt the player to guess the exact word using tiles from the rack, and the player will be asked to enter the word. 

If the move is valid, it will update the board. 
If the move is invalid, it will show an error message allowing the player to try again 3 times.

If the word is correct, the game will print that the player won, and the game will end.
If the player exceeds 3 trials, the game will print that the player lost, and the game will end.